+++ lib/public/components/common/newTable/columnModels/columnModel.js
@@ -1,0 +1,11 @@
+/**
+ * @interface
+ */
+export class ColumnModel {
+    /**
+     * @template T, V
+     * @param {T} row 
+     * @returns {V} data to be displayed in the column
+     */
+    getValue(row) {}
+}
