+++ lib/public/components/common/newTable/columnModels/passThroughColumnModel.js
@@ -1,0 +1,15 @@
+import { ColumnModel } from './columnModel.js'
+
+/**
+ * @implements {ColumnModel}
+ */
+export class PassThroughColumnmodel extends ColumnModel {
+    constructor(key) {
+        super()
+        this.key = key
+    }
+
+    getValue(coffeeData) {
+        return coffeeData[this.key]
+    }
+}
