+++ lib/public/components/common/newTable/structures/cellRenderer.js
@@ -2,6 +2,3 @@
  * @template V
  * @typedef {(V) => vnode} CellRenderer
  */
-
-// TODO does this V reference the output of ColumnModel.getValue ?
-// (defined in Coffee.js:6)
