+++ lib/public/components/common/newTable/table.js
@@ -9,8 +9,8 @@ import { h } from '/js/src/index.js';
  */
 const tableRow = (row, columns, getRowKey, getRowClasses, getRowConfig) => {
     const rowKey = getRowKey(row)
-    const rowClasses = getRowClasses(row)
-    const rowConfig = getRowConfig(row)
+    const rowClasses = getRowClasses ? getRowClasses(row) : ''
+    const rowConfig = getRowConfig ? getRowConfig(row) : {}
 
     return h(`tr.${rowClasses}`, {key: rowKey, ...rowConfig}, Object.entries(columns).map(([columnKey, column]) => {
         const cellData = column.model.getValue(row)
