+++ lib/public/views/Coffee/Coffee.js
@@ -1,31 +1,6 @@
 import { Observable } from '/js/src/index.js';
-
-/**
- * @interface
- */
-class ColumnModel {
-    /**
-     * @template T, V
-     * @param {T} row 
-     * @returns {V} data to be displayed in the column
-     */
-    getValue(row) {}
-}
-
-/**
- * @implements {ColumnModel}
- */
-export class PassThroughColumnmodel extends ColumnModel {
-    constructor(key) {
-        super()
-        this.key = key
-    }
-
-    getValue(coffeeData) {
-        return coffeeData[this.key]
-    }
-}
-
+import { ColumnModel } from '../../components/common/newTable/columnModels/columnModel.js';
+import { PassThroughColumnmodel } from '../../components/common/newTable/columnModels/passThroughColumnModel.js';
 /**
  * @implements {ColumnModel}
  */
