+++ lib/public/views/Environments/Overview/EnvironmentOverviewModel.js
@@ -14,7 +14,7 @@
 import { PaginationModel } from '../../../components/Pagination/PaginationModel.js';
 import { Observable, RemoteData } from '/js/src/index.js';
 import { getRemoteDataSlice } from '../../../utilities/fetch/getRemoteDataSlice.js';
-import { PassThroughColumnmodel } from '../../Coffee/Coffee.js';
+import { PassThroughColumnmodel } from '../../../components/common/newTable/columnModels/passThroughColumnModel.js';
 
 /**
  * Environment overview page model
