+++ lib/public/views/Environments/Overview/environmentOverviewComponent.js
@@ -90,4 +90,4 @@ const getEnvironmentTableRowKey = (row) => row.id
  * @param {Environment} row the environment corresponding to this row
  * @returns {Object} an object to be passed to h() with the row config
  */
-const getRowConfig = (row) => {onclick: console.log(row.id)}
+const getRowsConfig = (row) => {onclick: console.log(row.id)}
