+++ lib/public/views/Runs/Details/RunDetailsModel.js
@@ -21,7 +21,6 @@ import { TagPickerModel } from '../../../components/tag/TagPickerModel.js';
 import { TabbedPanelModel } from '../../../components/TabbedPanel/TabbedPanelModel.js';
 import { CollapsibleTreeNodeModel } from '../../../models/CollapsibleTreeNodeModel.js';
 import { detectorsProvider } from '../../../services/detectors/detectorsProvider.js';
-import { PassThroughColumnmodel } from '../../Coffee/Coffee.js';
 
 export const RUN_DETAILS_PANELS_KEYS = {
     LOGS: 'logs',
